split AppDelegate and MenuController
